accessible/tests/mochitest/attributes/test_obj_css.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        title="Accessibility CSS-based Object Attributes Test.">

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />

  <script type="application/javascript"
          src="../common.js" />
  <script type="application/javascript"
          src="../events.js" />
  <script type="application/javascript"
          src="../attributes.js" />

  <script type="application/javascript">
  <![CDATA[
    function doTest()
    {
      // CSS display
      testCSSAttrs("display_mozbox");
      testCSSAttrs("display_mozinlinebox");
      testCSSAttrs("display_mozgrid");
      testCSSAttrs("display_mozinlinegrid");
      testCSSAttrs("display_mozgridgroup");
      testCSSAttrs("display_mozgridline");
      testCSSAttrs("display_mozstack");
      testCSSAttrs("display_mozinlinestack");
      testCSSAttrs("display_mozdeck");
      testCSSAttrs("display_mozpopup");
      testCSSAttrs("display_mozgroupbox");

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  ]]>
  </script>

  <hbox flex="1" style="overflow: auto;">
  <body xmlns="http://www.w3.org/1999/xhtml">
    <a target="_blank"
       href="https://bugzilla.mozilla.org/show_bug.cgi?id=714579"
       title="Don't use GetComputedStyle for object attribute calculation">
      Mozilla Bug 714579
    </a><br/>

    <p id="display"></p>
    <div id="content" style="display: none">
    </div>
    <pre id="test">
    </pre>
  </body>

  <vbox id="display_mozbox" style="display: -moz-box;" role="img"/>
  <vbox id="display_mozinlinebox" style="display: -moz-inline-box;" role="img"/>
  <vbox id="display_mozgrid" style="display: -moz-grid;" role="img"/>
  <vbox id="display_mozinlinegrid" style="display: -moz-inline-grid;" role="img"/>
  <vbox id="display_mozgridgroup" style="display: -moz-grid-group;" role="img"/>
  <vbox id="display_mozgridline" style="display: -moz-grid-line;" role="img"/>
  <vbox id="display_mozstack" style="display: -moz-stack;" role="img"/>
  <vbox id="display_mozinlinestack" style="display: -moz-inline-stack;" role="img"/>
  <vbox id="display_mozdeck" style="display: -moz-deck;" role="img"/>
  <vbox id="display_mozpopup" style="display: -moz-popup;" role="img"/>
  <vbox id="display_mozgroupbox" style="display: -moz-groupbox;" role="img"/>

  </hbox>
</window>
